refactor(gui): split keyboard-shortcut system into MainWindow_Shortcuts.cpp (#3351 Phase 1c)
Fourth PR of the #3351 monolith-decomposition series. Pure code motion —
no behavior change, no MainWindow.h changes.

Moves the complete keyboard-shortcut system out of MainWindow.cpp into
the new MainWindow_Shortcuts.cpp (1,087 lines):

- The shortcut-state definitions (s_keyboardShortcutsEnabled,
  s_sliderShortcutLeaseActive) and accessors (textInputCaptured,
  shortcutInputCaptured, shortcutGuard, leaseHolderBusy) — declared in
  MainWindowShortcutState.h since Phase 1b, owned here as of this phase,
  completing the handoff that header promised.
- registerShortcutActions() (559 lines — the full action table).
- The slider shortcut lease (#745): beginSliderShortcutLease,
  renewSliderShortcutLease, releaseSliderShortcutLease, plus
  MainWindow::eventFilter() (319 lines) which drives the lease and the
  app-quit interception.
- handleCwMomentaryShortcut() and the trivial keyPress/keyRelease
  overrides.

Supporting moves, called out for review:

- kSliderShortcutLeaseMs moved with its only user.
- panCountForLayoutId promoted from MainWindow.cpp's anonymous namespace
  to MainWindowHelpers.{h,cpp} — registerShortcutActions (moved) and two
  MainWindow.cpp call sites now share it across TUs. Pure map lookup;
  fits the helpers file's no-state rule.
- isCwMomentaryActionId NOT moved — its remaining caller is the MIDI
  dispatch path still in MainWindow.cpp; it stays file-static there.

Breadcrumb comments at all six excision sites.

MainWindow.cpp: 16,261 → 15,239 lines (series: 19,474 → 15,239, −21.7%).
Verified: full build clean, all 33 ctest suites pass (incl. the new
radio_discovery_test from main), a11y lint clean.

Refs #3351. Principle XI.
